+++ doc/CHANGES.rst
@@ -5,6 +5,12 @@ This file contains change information for the current Zope release.
 Change information for previous versions of Zope can be found at
 http://docs.zope.org/zope2/
 
+2.13.30 (unreleased)
+--------------------
+
+- Nothing changed yet.
+
+
 2.13.29 (2019-02-09)
 --------------------